+++ packages/address/src/index.ts
@@ -1,4 +1,4 @@
-export * from './Mock/MockAddress';
+export * from './bb/MockAddress';
 export * from './Address';
 export * from './ImmutableAddress';
 export * from './MutableAddress';
