+++ packages/sequence/src/index.ts
@@ -1,4 +1,4 @@
-export * from './Mock/MockSequence';
+export * from './bb/MockSequence';
 export * from './ImmutableSequence';
 export * from './MutableSequence';
 export * from './ReadonlySequence';
