+++ packages/tree/src/ClosureTable/__tests__/ClosureTable.spec.ts
@@ -1,12 +1,12 @@
 import { MockAddress, ReadonlyAddress } from '@jamashita/lluvia-address';
 import { MockProject, Project } from '@jamashita/lluvia-project';
 import { ReadonlySequence } from '@jamashita/lluvia-sequence';
-import { MockTreeID } from '../../Mock/MockTreeID';
+import { MockTreeID } from '../../bb/MockTreeID';
+import { MockClosureTable } from '../bb/MockClosureTable';
+import { MockClosureTableHierarchies } from '../bb/MockClosureTableHierarchies';
+import { MockClosureTableHierarchy } from '../bb/MockClosureTableHierarchy';
 import { ClosureTable } from '../ClosureTable';
 import { ClosureTableHierarchies } from '../ClosureTableHierarchies';
-import { MockClosureTable } from '../Mock/MockClosureTable';
-import { MockClosureTableHierarchies } from '../Mock/MockClosureTableHierarchies';
-import { MockClosureTableHierarchy } from '../Mock/MockClosureTableHierarchy';
 
 describe('ClosureTable', () => {
   describe('of', () => {
