+++ packages/tree/src/ClosureTable/index.ts
@@ -1,4 +1,4 @@
-export * from './Mock/index';
+export * from './bb/index';
 export * from './ClosureTable';
 export * from './ClosureTableHierarchies';
 export * from './ClosureTableHierarchy';
