+++ packages/tree/src/TreeNode/index.ts
@@ -1,4 +1,4 @@
-export * from './Mock/MockTreeNode';
+export * from './bb/MockTreeNode';
 export * from './ATreeNode';
 export * from './ReadonlyTreeNode';
 export * from './SerializableTreeNode';
