+++ packages/tree/src/__tests__/SerializableTrees.spec.ts
@@ -1,6 +1,6 @@
 import { ImmutableAddress } from '@jamashita/lluvia-address';
-import { MockTreeID } from '../Mock/MockTreeID';
-import { MockTreeObject } from '../Mock/MockTreeObject';
+import { MockTreeID } from '../bb/MockTreeID';
+import { MockTreeObject } from '../bb/MockTreeObject';
 import { SerializableTree } from '../SerializableTree';
 import { SerializableTrees } from '../SerializableTrees';
 import { SerializableTreeNode } from '../TreeNode/SerializableTreeNode';
