+++ packages/tree/src/__tests__/StructurableTrees.spec.ts
@@ -4,11 +4,11 @@ import { ImmutableProject } from '@jamashita/lluvia-project';
 import { ImmutableSequence } from '@jamashita/lluvia-sequence';
 import { ClosureTable } from '../ClosureTable/ClosureTable';
 import { ClosureTableHierarchies } from '../ClosureTable/ClosureTableHierarchies';
-import { MockClosureTableHierarchies } from '../ClosureTable/Mock/MockClosureTableHierarchies';
-import { MockClosureTableHierarchy } from '../ClosureTable/Mock/MockClosureTableHierarchy';
+import { MockClosureTableHierarchies } from '../ClosureTable/bb/MockClosureTableHierarchies';
+import { MockClosureTableHierarchy } from '../ClosureTable/bb/MockClosureTableHierarchy';
 import { TreeError } from '../Error/TreeError';
-import { MockTreeID } from '../Mock/MockTreeID';
-import { MockTreeObject } from '../Mock/MockTreeObject';
+import { MockTreeID } from '../bb/MockTreeID';
+import { MockTreeObject } from '../bb/MockTreeObject';
 import { StructurableTree } from '../StructurableTree';
 import { StructurableTrees } from '../StructurableTrees';
 import { StructurableTreeNode } from '../TreeNode/StructurableTreeNode';
