+++ packages/tree/src/__tests__/Tree.spec.ts
@@ -1,8 +1,8 @@
 import { ImmutableAddress } from '@jamashita/lluvia-address';
-import { MockTree } from '../Mock/MockTree';
-import { MockTreeID } from '../Mock/MockTreeID';
-import { MockTreeObject } from '../Mock/MockTreeObject';
-import { MockTreeNode } from '../TreeNode/Mock/MockTreeNode';
+import { MockTree } from '../bb/MockTree';
+import { MockTreeID } from '../bb/MockTreeID';
+import { MockTreeObject } from '../bb/MockTreeObject';
+import { MockTreeNode } from '../TreeNode/bb/MockTreeNode';
 
 describe('Tree', () => {
   describe('equals', () => {
