+++ packages/tree/src/bb/MockTrees.ts
@@ -1,7 +1,7 @@
 import { MutableProject, ReadonlyProject } from '@jamashita/lluvia-project';
 import { ATrees } from '../ATrees';
 import { TreeID } from '../TreeID';
-import { MockTreeNode } from '../TreeNode/Mock/MockTreeNode';
+import { MockTreeNode } from '../TreeNode/bb/MockTreeNode';
 import { MockTree } from './MockTree';
 import { MockTreeObject } from './MockTreeObject';
 
@@ -10,7 +10,7 @@ export class MockTrees<K extends TreeID, V extends MockTreeObject<K>> extends AT
     super(MutableProject.of<K, MockTree<K, V>>(trees));
   }
 
-  public add(tree: MockTree<K, V>): MockTrees<K, V> {
+  public add(tree: MockTree<K, V>): this {
     this.trees.set(tree.getRoot().getValue().getTreeID(), tree);
 
     return this;
